[project @ 2002-07-04 13:33:24 by simonmar] Sigh, enabling _POSIX_PTHREAD_SEMANTICS causes the Solaris header files to drop several silly little C functions into the source code. We don't want these duplicated in every Haskell-compiled object.
So plan B: just define _POSIX_PTHREAD_SEMANTICS in dirUtils.c where it
is needed to make readdir_r work.  I've checked, and readdir_r is the
only function we use that is affected by _POSIX_PTHREAD_SEMANTICS.  If
we ever use any more of these functions, then we'll have to be careful
to give them a C wrapper.