[project @ 2002-09-25 15:24:07 by simonmar] Re-instate the checking for the values of errno constants at configure time.  The problem with doing it using foreign calls is simply that this tickles a bad case in the code gen machinery, which in this case results in an extra 10-20k of goop ending up in pretty much every binary, and it impacts GC performance too.
This has some portability implications, but the situation is no worse
than before.  To reliably cross-compile for a new platform you need to
build a set of .hc files for the libraries using a config.h generated on
the *target* machine.  (at some point we'll formalise the
cross-compilation story, but that's another thing on the todo list...)