[project @ 2003-04-09 08:18:13 by simonpj] -------------------------------------       Fix the lexer so that it does the right thing 	for floating point and hexadecimal numbers 	-------------------------------------
This fix requires the new (<++) combinator that Koen has
added to the ReadP library.

NB: we lex "0x9fw" as [("0x9f", "w")]; that is, we stop as soon
as we get a legal thing (maximally munching) even if white space
does not follow.  I'm not totally certain this is the right thing,
but it's much better than it was before.