Changing delta to 3 in Data.Set.
Only possible values are 3 and 4. The value 3 has much faster inserts,
value 4 slightly faster deletes, so choosing 3.

Also changed the inequalities to rebalance only when one subtree
is _strictly_ larger than delta * the other one, to mimic the behaviour
from the proof (both from the Adams' and from the one to come).