Remove most INLINE from Map, Set, IntMap and IntSet.
Because of a code bloat the INLINEs cause, remove most of
them. The only INLINEs left are the following:
- only in Set and Map, because in IntMap and IntSet the specialization
  does not help
- only on functions which need Ord
- only on 'small' functions, namely member, notMember, lookup*,
  insert*, delete*, adjust*, alter*, update*

All other functions of Map, Set, IntMap and IntSet are marked INLINABLE,
even if they are recursive.

The INLINEs left are only a short-term solution. In the long run the
auto-specialization of INLINABLE methods seems a good way (maybe
SPECIALIZABLE).