Change the worker/wrapper to explicitly pass arguments.
As the benchmarking showed, it is not a good idea to create
closures in the worker/wrapper transformation, as the captured
arguments of the enclosing function have to be allocated on the
heap. It is better to explicitly pass the arguments on the stack.
This saves memory and add no time penalty if the arguments are
the first arguments of recursive function (GHC refrains from
needless copying).

The worker is often strict in some arguments. I did not want
to use BangPatterns, so I used macros to indicate strictness.
If majority thinks BangPatters are fine, I will gladly change it.