Changing INLINE pragmas.
The internal functions that need to be inlined are marked INLINE
(bit fiddling in IntMap/IntSet, empty, singleton, bin).

Aslo if INLINABLE is available, use it for all exported functions.
The functions like insert that were INLINE because of specialization
issues are now INLINE only in the case of INLINABLE absence.