Add all fold variants.
Add foldr, foldl, foldr', foldl' to Data.*Set and Data.*Map.
The fold function is in Legacy section and said to be deprecated
and removed in the future.

Also add foldrWithKey, foldlWithKey, foldrWithKey', foldlWithKey
to Data.*Map. The foldWithKey function is in Legacy section and said
to be deprecated and removed in the future.