Improve performance of folds.
We use wordsize-dependent implementation for GHC, for both
32-bit and 64-bit architectures.

It is based on fast constant-time implementation of indexOfTheOnlyBit,
which computes index of the only bit set in a word, suggested
by Edward Kmett.

Using that we can enumerate indexes of 1 bits, in the order from
LSB to MSB. That results in fast foldl implementations.

Foldr implementations bit-reverse the word and then iterate from the LSB
to MSB using accumulator. That is faster then either not using
accumulator or iterating from MSB to LSB.