Improvements of test-suites compilation.
Previously the test-suites could be compiled only when a special
testing version of the library was built using -ftesting flag.
As commented by Duncan, that is bad -- users which enable testing
by default would install the testing version of library.

Now the test-suites do not link against the whole containers,
but compiles the necessary modules separately, with special testing
flags.