Improve {Map, IntMap}.keysSet.
The keysSet method is now implemented using the exported constructors of
Data.{Set, IntSet}.Base.

The implementation of Map.keysSet is trivial, as Set and Map use same
tree structure.

The implementation of IntMap.keysSet is slightly complicated, because of
the dense representation of IntSet, where several last levels of the
tree are flatten into a bitmap.