Remove references to Report and FFI from LICENSE
The LICENSE claimed that the source code of containers comes from three
sources:
* from GHC project, licensed under GHC license
* from Haskell Report, licensed under Haskell Report license
* from FFI, licensed under FFT license
It did not specify which fragments of code came from where. That caused
confusion among the containers users.

Me (Milan Straka) and Johan Tibell believe that in the containers source
there is no code derived from Haskell Report, and that in the containers
source there is no code derived from FFI, so I removed the references to
Haskell Report and FFI from the LICENSE.

Now the LICENSE file contains only the GHC license, without any
comments.