tree: Fix imports for the Applicative/Monad change
Due to various problems with orphans and cycles in base, while
implementing the Applicative/Monad Proposal, Alternative joined
MonadPlus in Control.Monad.

A knock-on effect of this is that Control.Monad now exports 'empty',
which conflicts with Data.Sequence in this case. Luckily the fix is
actually quite easy: just restrict the imports to liftM, since that's
all we use anyway.

Signed-off-by: Austin Seipp <aseipp@pobox.com>