Speed up Data.Sequence.splitAt and zipWith
Previously, `splitAt` returned a lazy pair, and was pretty much
lazy throughout. Now it's about as strict as it's allowed to be.

Avoid allocating anything extra when splitting at or before the
beginning of a sequence.

Fix misplaced bang annotation (my fault).

Hand-inline and reduce the splitting of the top of the tree. This
avoids a separate cons step at the end, saving some time when
splitting small sequences. This is particularly helpful for
zip performance.

Note: `zip` performance is, for some reason, rather sensitive to
how the inliner handles `deepL`. I don't know why. I also don't
really know how to fix it. Splitting in general is rather complicated
and hard to follow. It would be nice to fix that.

Rewrite `take` and `drop` so they don't build the parts of the
tree they don't actually use. Previously, they were written
using `splitAt`.

Use custom left and right views for `FingerTree`s. This was a side
effect of something I thought I wanted to do, but it's a good idea
anyway.