Stop fearing the middle
A couple years ago I thought it *must* be a good idea to ever look down
the middle if it could possibly be avoided, but I never had any
benchmarks to support it, and it seems pretty silly in retrospect.
Reverting pending some numbers demonstrating it's really the way to go.