Rename merge modules
I think it's more consistent with the rest of the API to name
them `Data.Map.Merge.Lazy` and `Data.Map.Merge.Strict`. This also
gives us the option to add further merge-related modules in the
`Merge` hierarchy. The original names still work for now, but they
are deprecated and hidden from Haddock.