Make min and max views more eager
* Make min and max views in `Data.IntMap` more
eager, and structure them to reveal `Nothing` or `Just` as soon as
possible. This lets GHC erase `Maybe` constructors and unbox tuples, at
least when it's in a good mood.

* Add `lookupMin` and `lookupMax` to `Data.IntMap`
as total alternatives to `findMin` and `findMax`.

* Add `!?` to `Data.IntMap` as a total alternative to `!`.

* Set the fixity of `Data.IntMap.!` to match that of `Data.Map.!`.

* Make more update functions total. Historically, many have thrown
errors when given missing keys. Now, we prefer to return the structures
unchanged.