Add another way to build benchmarks
Building benchmarks while using `containers` as a library never worked
for me, because some of the dependencies are usually always installed
system-wide (`binary`, an older version of `containers` etc.).

So this package instead uses the library files as source files (instead
of using the library as a library dependency). This means that we link
whatever `containers` library available with dependencies like `binary`
and `criterion`, but we use `containers` source files from the parent
directory when compiling benchmark programs. In my case this always
works.

One disadvantage is that we re-compile `containers` source files again
and again for every single benchmark, so compile times suffer.