Add validity checks for IntSet and IntMap.
The checks codify the invariants laid out in the IntSet and IntMap comments. We
add validity checks for constructed IntSets and IntMaps as well as union,
intersection, and difference operations on them.

As it turns out, the mask being a power of 2 checks do not hold for the IntMap
implementation; this may be something to investigate.