Improve inference for IsString instance for Seq
When I originally added an `IsString` instance for `Seq`, we
conservatively used

```haskell
instance IsString (Seq Char)
```

to allow for the possibility that we might eventually want to write
additional instances. But the result is very bad inference and
a discrepency with the instance for lists. So let's not do that.