Introduction/tutorial doc improvements. (#512)
Changes:
--------

- Update titles so that package name is visible in browser tabs
- Fix typo in sequence docs
- Use https://github.com/m-renaud/haddock-autolink
- Update theme (dark grey with green) for desktop and mobile

Why use a submodule?
--------------------

Improvements to the autolinker don't need to take place in the containers repo. Also, if other projects end up using it the code doesn't get out of sync (not everyone needs to copy/paste the Sphinx extension).

If you're not modifying the docs you don't need to pull down all the Python code. If you do want to build the container docs locally you need to pull down the submodule code (see the CONTRIBUTING.md file for exact commands to run).

Read the docs natively supports submodules so this has no visible effect on the docs.

[ci skip]