Make Data.Graph use a bit vector
Use a bit vector in `Data.Graph` instead of an array of
`Bool`. This reduces the size of the array from the number
of words the graph has vertices to the number of vertices
divided by the word size.

Is this a good idea? I don't know for sure. If we *don't* do
this, we should *almost certainly* make it use an unboxed array
rather than a boxed one; we don't need the lazness, and we surely
don't need the garbage collector running through arrays of Bool!