Unbox the `SetM` array
`Data.Graph` uses a mutable array to track which vertices have
been visited. This was implemented using an `STArray s Vertex Bool`.
This is obviously bad, because the garbage collector will have to
trace the array if it's alive during a collection. Using an unboxed
array instead should always be slightly better.

It may be better to go further and use a bit array to save space,
but I'm not entirely comfortable making that change until we have
some tests for `Data.Graph`--any mistakes there will likely be
subtle.