Disable deprecated things
Disable all deprecated functions by making them produce type errors when used.
This is an experimental removal approach.  It may or may not really be a good
idea. The benefit is that unlike deprecation it *forces* people to make a
change, and unlike immediate removal it tells them what change to make.

With GHC 8.0 and above, the type errors are pretty. With earlier
versions of GHC, they're gross. With a hypothetical non-GHC compiler, the
functions are just removed.