Add examples for foldTree
The two existing examples could also be implemented using the `Foldable` instance of `Tree`, although I realize that the `Foldable` `maximum` implementation is partial, while the example using `foldTree` is total.

The two additional examples are operations that, as far as I can tell, can't be implemented with `Foldable`, but only with the full catamorphism. This, hopefully, might be compelling examples.