indexOfTheOnlyBit -> countTrailingZeros
CPUs can do this very quickly these days and the implementation is
readily available since base-4.8. There are also fewer
assumptions (only set bit) which is nice.

There is a small change in behaviour when `0` is fed in: previously
we'd get `63` for a 64-bit word, now we get `64`. We don't try to
catch this case: this fact wasn't used anywhere and the internal
invariant is that a bit is always set. If necessary, it's very easy to
replicate the old behaviour at some performance cost.