Split out Set.insertNE
We cannot do pointer equality very easily on interior nodes, but
thankfully we don't need to! It's decided once at the leaf whether a new
set is needed, so we can propagate that decision back with Maybe.

If GHC is smart enough to get rid of the Maybe entirely with two
continuations, this might be even faster.