Improve fromList for IntSet
Make `fromList` for `IntSet` better for partially sorted input.
Performance seems to be similar to the old implementation for
random input, but nearly as fast as `fromDistinctAscList` for
sorted or reverse sorted input. There are pathological cases where
the new implementation is significantly but not horribly slower than
the old. In particular, I noticed that

```haskell
iterate (\n -> (n ^ 2) `rem` (2^32-1)) 20
```

is pretty bad for the new implementation for some reason.