Revert union, difference, etc. complexity changes
As discussed in #870, the complexities for union, difference, etc. on
Set and Map were changed in #830 to fix some partiality in the
expressions, but along the way new partiality was introduced, and useful
special cases like m = 1 get incorrect complexity values from the new
formulas.

The original formula was as stated in the original paper:
  https://dl.acm.org/doi/10.1145/322123.322127
and this holds for 0 < m <= n, which seems sufficient to me. (The m=0
case is excluded, but for m=0 nothing needs to be done anyway. This
contrary to the m=1 case, in which useful work with very specific
complexity (namely, O(log(n))) needs to be done.)

This commit reverts all occurrences of the modified complexity formula
back to the original one.