Enable -fworker-wrapper-cbv on ghc-9.4
This flag allows for some significant perf improvements and the
downsides don't apply to containers.

In particular this flag can cause rules to not fire *if* the relevant
functions don't have NOINLINE pragmas. However the relevant functions
in containers seem to have such a pragma so there should be no downside.